Work around a git issue
The git version (2.17) in the Ubuntu repos is throwoing a TLS error
when cloning the upstream tmux repo. It's not related to the work proxy
cert since I've already added the cert to the image I am using.

Installing a new version of git (2.33) still shows the error but it
does clone the repo and the return code is 0.